requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
langchain==0.1.13
langchain-openai==0.1.1
python-dotenv==1.0.1
boto3>=1.34.0
fastapi
uvicorn
python-multipart
pdfplumber
edge-tts
openai-whisper
moviepy==1.0.3
